Add middleware to validate checksum
This checksum is sent in a X-Checksum header and it is calculated using
the HMAC-SHA1 of the json string of an object including url, query parameters and body
using the key "all your base are belong to us"

Currently not enabled because  it's cumbersome to calculate this sum when
developing the api. It will be enabled in production.

example: "{ "url": "/users", "query": "{ "s": "example" }, "body": "{}" }"